% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedDataTable-package.R
\docType{package}
\name{LinkedDataTable-package}
\alias{LinkedDataTable}
\alias{LinkedDataTable-package}
\title{LinkedDataTable: Operators for Collections of 'data.table' Objects}
\description{
Provides the class 'LinkedDataTable' for performing operations on collections of 'data.table' objects which are related by a common column.
}
\keyword{internal}
